feat: dao overview screen
