+++ src/utils/treasury.ts
@@ -1,0 +1,12 @@
+import BigNumber from 'bignumber.js'
+import { TTreasuryCoinModel } from 'pages/CurrentEntity/Treasury/DAOTreasury/Accounts'
+
+export function getTotalUSDvalueFromTreasuryCoins(coins: { [denom: string]: TTreasuryCoinModel }): string {
+  return (
+    Object.values(coins).reduce(
+      (pre, cur) =>
+        new BigNumber(pre).plus(new BigNumber(cur.balance).times(new BigNumber(cur.lastPriceUsd))).toString(),
+      '0',
+    ) ?? '0'
+  )
+}
